fix parse line;

+++ src/main/java/Asessor.java
@@ -1,11 +1,11 @@
 public class Asessor {
     String login;
     Integer uid;
-    Short docid;
+    Integer docid;
     Short jud;
     Short cjud;
 
-    public Asessor(String login, Integer uid, Short docid, Short jud, Short cjud) {
+    public Asessor(String login, Integer uid, Integer docid, Short jud, Short cjud) {
         this.login = login;
         this.uid = uid;
         this.docid = docid;
